comm: move cert extraction to common/util
There's already code over there to extract information from the gRPC
context so it's as good a home as any.
