+++ common/util/net_test.go
@@ -8,9 +8,13 @@ package util
 
 import (
 	"context"
+	"crypto/sha256"
+	"crypto/tls"
+	"crypto/x509"
 	"testing"
 
 	"github.com/stretchr/testify/require"
+	"google.golang.org/grpc/credentials"
 	"google.golang.org/grpc/peer"
 )
 
@@ -32,4 +36,38 @@ func TestExtractAddress(t *testing.T) {
 		Addr: &addr{},
 	})
 	require.Equal(t, "1.2.3.4:5000", ExtractRemoteAddress(ctx))
+}
+
+func TestExtractCertificateHashFromContext(t *testing.T) {
+	require.Nil(t, ExtractCertificateHashFromContext(context.Background()))
+
+	p := &peer.Peer{}
+	ctx := peer.NewContext(context.Background(), p)
+	require.Nil(t, ExtractCertificateHashFromContext(ctx))
+
+	p.AuthInfo = &nonTLSConnection{}
+	ctx = peer.NewContext(context.Background(), p)
+	require.Nil(t, ExtractCertificateHashFromContext(ctx))
+
+	p.AuthInfo = credentials.TLSInfo{}
+	ctx = peer.NewContext(context.Background(), p)
+	require.Nil(t, ExtractCertificateHashFromContext(ctx))
+
+	p.AuthInfo = credentials.TLSInfo{
+		State: tls.ConnectionState{
+			PeerCertificates: []*x509.Certificate{
+				{Raw: []byte{1, 2, 3}},
+			},
+		},
+	}
+	ctx = peer.NewContext(context.Background(), p)
+	h := sha256.New()
+	h.Write([]byte{1, 2, 3})
+	require.Equal(t, h.Sum(nil), ExtractCertificateHashFromContext(ctx))
+}
+
+type nonTLSConnection struct{}
+
+func (*nonTLSConnection) AuthType() string {
+	return ""
 }
